+++ include/clap/host.h
@@ -15,9 +15,12 @@ typedef struct clap_host {
    const char *name;    // eg: "Bitwig Studio"
    const char *vendor;  // eg: "Bitwig GmbH"
    const char *url;     // eg: "https://bitwig.com"
-   const char *version; // eg: "4.3"
+   const char *version; // eg: "4.3", see plugin.h for advice on how to format the version
 
    // Query an extension.
+   // The returned pointer is owned by the host.
+   // It is forbidden to call it before plugin->init().
+   // You can call it within plugin->init() call, and after.
    // [thread-safe]
    const void *(CLAP_ABI *get_extension)(const struct clap_host *host, const char *extension_id);